Fix broken ament_install_python function
The upstream package had copied this from catkin_virtualenv and
updated it for ament. But the templates were not copied, and some
of the cmake variables were broken and not correct.

This patch fixes this by copying the install template, fixing the
broken variables, and removing the "devel" logic as this doesn't
apply to ROS2 builds.